refactor(agent-client): move test-only permission overrides to agent-testing
The permission-override mechanism (overridePermissions field, the
overrideCollectionPermission / overrideActionPermission / clearPermissionOverride
methods, and the *PermissionsOverride types) was only ever used by agent-testing,
yet it lived on the production RemoteAgentClient base class — exposing no-op
methods on the prod client API.

Move it to TestableAgentBase (agent-testing), the sole consumer:
- agent-client: drop the field/methods/types from RemoteAgentClient and the dead
  overridePermissions param from createRemoteAgentClient.
- agent-testing: TestableAgentBase now holds overridePermissions + the override
  methods; the types live in a new permission-overrides module used by the
  sandbox and createAgentTestClient.

BREAKING CHANGE: @forestadmin/agent-client no longer exports
createRemoteAgentClient's overridePermissions option nor the
PermissionsOverride / CollectionPermissionsOverride / SmartActionPermissionsOverride
types (test-only; use @forestadmin/agent-testing).
